+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Gradle Grammar-Kit Plugin Changelog
 
 ## [Unreleased]
+- Set minimum supported Gradle version from `6.7.1` to `6.8`
 
 ## [2021.2.2]
 - Remove redundant `bomConfiguration`